blog-generator/services/requirements.txt
# Core FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6

# OpenAI and environment management
openai>=1.3.0
python-dotenv>=1.0.0

# HTTP requests and async support
requests>=2.31.0
httpx>=0.25.0
aiofiles>=23.2.0

# Additional utilities that might be needed
typing-extensions>=4.8.0
anyio>=4.0.0

# JSON and data processing (usually built-in but sometimes needed explicitly)
# These are typically included in Python standard library but adding for safety
email-validator>=2.1.0